pipeline-manager: split pipeline-specific DB calls into a separate module
