+++ crates/pipeline_manager/proptest-regressions/db/test.txt
@@ -34,3 +34,4 @@ cc d7f951afedd9798076cc9b3b0e141de9e23766af2eca698eb58c8641f91295f5 # shrinks to
 cc 10a2e3004b09211bfd73fa87d7972814911f65890e5fbf0c6039ac2a52e2f300 # shrinks to [NewProgram(TenantId(01000000-0000-0000-0000-000000000000), 03000000-0000-0000-0000-000000000000, "", "", ""), NewPipeline(TenantId(01000000-0000-0000-0000-000000000000), 01000000-0000-0000-0000-000000000000, None, "", "", RuntimeConfig { workers: 0, cpu_profiler: false, min_batch_size_records: 0, max_buffering_delay_usecs: 0, resources: ResourceConfig { cpu_cores_min: None, cpu_cores_max: None, memory_mb_min: None, memory_mb_max: None, storage_mb_max: None } }, None), GetPipelineDescrByName(TenantId(01000000-0000-0000-0000-000000000000), "")]
 cc d14f3926aacea803a6d664eaf6f915042033819fca0050a8d0bb93d2014b1b4a # shrinks to [NewProgram(TenantId(01000000-0000-0000-0000-000000000000), 01000000-0000-0000-0000-000000000000, "", "", ""), SetProgramStatusGuarded(TenantId(01000000-0000-0000-0000-000000000000), ProgramId(01000000-0000-0000-0000-000000000000), Version(1), Pending), SetProgramStatusGuarded(TenantId(01000000-0000-0000-0000-000000000000), ProgramId(01000000-0000-0000-0000-000000000000), Version(1), None), AllPrograms]
 cc 9add742fd6dd4d8a6f2408504ce4e0c63bbef5153259c613d33719c8bb448ad7 # shrinks to [NewProgram(TenantId(01000000-0000-0000-0000-000000000000), 01000000-0000-0000-0000-000000000000, "", "", ""), SetProgramSchema(TenantId(01000000-0000-0000-0000-000000000000), ProgramId(01000000-0000-0000-0000-000000000000), ProgramSchema { inputs: [], outputs: [] }), SetProgramStatusGuarded(TenantId(01000000-0000-0000-0000-000000000000), ProgramId(01000000-0000-0000-0000-000000000000), Version(1), Pending), AllPrograms]
+cc 9837ee95836fb9c469e8b728897ba81414d7e4904177202aa89c4de5f071a26a # shrinks to [NewProgram(TenantId(03000000-0000-0000-0000-000000000000), 03000000-0000-0000-0000-000000000000, "", "", ""), UpdateProgram(TenantId(03000000-0000-0000-0000-000000000000), ProgramId(03000000-0000-0000-0000-000000000000), None, None, Some("ௐ"), None), NewProgram(TenantId(01000000-0000-0000-0000-000000000000), 01000000-0000-0000-0000-000000000000, "", "", ""), UpdateProgram(TenantId(01000000-0000-0000-0000-000000000000), ProgramId(01000000-0000-0000-0000-000000000000), None, None, Some("ⷘ"), None), SetProgramStatusGuarded(TenantId(01000000-0000-0000-0000-000000000000), ProgramId(01000000-0000-0000-0000-000000000000), Version(2), Pending), SetProgramStatusGuarded(TenantId(03000000-0000-0000-0000-000000000000), ProgramId(03000000-0000-0000-0000-000000000000), Version(2), Pending), NextJob]
